LayoutTests/http/tests/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas-expected.txt

CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: There are too many active WebGL contexts on this page, the oldest context will be lost.
CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 268435456).

PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="srgb";snapshotType="toDataURL"
FAIL :onscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="srgb";snapshotType="toBlob" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                               1,0,0                                0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                         00 00 00 ff                          00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000     0.00000,0.00000,0.00000,1.00000      0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.380392,1.00000 0.168627,0.392157,0.0980392,1.00000 0.364706,0.0549020,0.0274510,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                           0,0,0,255                            0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                       0,0,97,255                       43,100,25,255                          93,14,7,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                 0,0,-97,0                      -43,-100,-25,0                         -93,-14,-7,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
 Reached unreachable code
PASS :onscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="srgb";snapshotType="imageBitmap"
FAIL :onscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";snapshotType="toDataURL" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.419608,1.00000 0.00000,0.407843,0.00000,1.00000 0.439216,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,107,255                      0,104,0,255                      112,0,0,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-107,0                       0,-104,0,0                       -112,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
 Reached unreachable code
PASS :onscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";snapshotType="toBlob"
PASS :onscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :onscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";snapshotType="toDataURL"
PASS :onscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";snapshotType="toBlob"
FAIL :onscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";snapshotType="imageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            1,0,0                            0,1,0                            1,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 1.00000,1.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,255,0,102                      255,0,0,102                    255,255,0,102 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,-255,0,-102                    -255,0,0,-102                 -255,-255,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                               1,0,0                               0,1,0                             1,1,0 ...
      act. texel bytes (little-endian) ==      0x:                         00 00 00 00                         00 00 00 00                       00 00 00 00 ...
                           act. colors == R,G,B,A:     0.00000,0.00000,0.00000,0.00000     0.00000,0.00000,0.00000,0.00000   0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.462745,0.980392,0.294118,0.400000 0.921569,0.196078,0.137255,0.400000 1.00000,1.00000,0.333333,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                             0,0,0,0                             0,0,0,0                           0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      118,250,75,102                       235,50,35,102                    255,255,85,102 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                           -118,-250,-75,-102                   -235,-50,-35,-102                -255,-255,-85,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
 Reached unreachable code
FAIL :onscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="toDataURL" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,1.00000,0.400000 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,255,102                      0,255,0,102                      255,0,0,102 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,0,-255,-102                    0,-255,0,-102                    -255,0,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,1.00000,0.400000 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,255,102                      0,255,0,102                      255,0,0,102 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,0,-255,-102                    0,-255,0,-102                    -255,0,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
 Reached unreachable code
FAIL :onscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="toBlob" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,1,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,1,0                            1,1,0
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000
                           exp. colors == R,G,B,A: 1.00000,0.00000,0.00000,0.400000 1.00000,1.00000,0.00000,0.400000
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0
            exp. normal-ULPs-from-zero == R,G,B,A:                      255,0,0,102                    255,255,0,102
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             -255,0,0,-102                 -255,-255,0,-102
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,1,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,1,0                            1,1,0
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000
                           exp. colors == R,G,B,A: 1.00000,0.00000,0.00000,0.400000 1.00000,1.00000,0.00000,0.400000
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0
            exp. normal-ULPs-from-zero == R,G,B,A:                      255,0,0,102                    255,255,0,102
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             -255,0,0,-102                 -255,-255,0,-102
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:303:19
 Reached unreachable code
PASS :onscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="opaque";webgl="webgl";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="opaque";webgl="webgl";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="opaque";webgl="webgl2";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="opaque";webgl="webgl2";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="premultiplied";webgl="webgl";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="premultiplied";webgl="webgl";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="premultiplied";webgl="webgl2";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="bgra8unorm";alphaMode="premultiplied";webgl="webgl2";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="opaque";webgl="webgl";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="opaque";webgl="webgl";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="opaque";webgl="webgl2";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="opaque";webgl="webgl2";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="premultiplied";webgl="webgl";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="premultiplied";webgl="webgl";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="premultiplied";webgl="webgl2";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba8unorm";alphaMode="premultiplied";webgl="webgl2";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="opaque";webgl="webgl";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="opaque";webgl="webgl";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="opaque";webgl="webgl2";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="opaque";webgl="webgl2";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="premultiplied";webgl="webgl";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="premultiplied";webgl="webgl";upload="texSubImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="premultiplied";webgl="webgl2";upload="texImage2D"
PASS :onscreenCanvas,uploadToWebGL:format="rgba16float";alphaMode="premultiplied";webgl="webgl2";upload="texSubImage2D"
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                               1,0,0                                0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                         00 00 00 ff                          00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000     0.00000,0.00000,0.00000,1.00000      0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.380392,1.00000 0.168627,0.392157,0.0980392,1.00000 0.364706,0.0549020,0.0274510,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                           0,0,0,255                            0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                       0,0,97,255                       43,100,25,255                          93,14,7,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                 0,0,-97,0                      -43,-100,-25,0                         -93,-14,-7,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.419608,1.00000 0.00000,0.407843,0.00000,1.00000 0.439216,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,107,255                      0,104,0,255                      112,0,0,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-107,0                       0,-104,0,0                       -112,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 1,1,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            1,1,0
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000
                           exp. colors == R,G,B,A: 1.00000,1.00000,0.00000,0.400000
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0
            exp. normal-ULPs-from-zero == R,G,B,A:                    255,255,0,102
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                          -255,-255,0,-102
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 1,1,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                             1,1,0
      act. texel bytes (little-endian) ==      0x:                       00 00 00 00
                           act. colors == R,G,B,A:   0.00000,0.00000,0.00000,0.00000
                           exp. colors == R,G,B,A: 1.00000,1.00000,0.333333,0.400000
            act. normal-ULPs-from-zero == R,G,B,A:                           0,0,0,0
            exp. normal-ULPs-from-zero == R,G,B,A:                    255,255,85,102
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                         -255,-255,-85,-102
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="convertToBlob"
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="transferToImageBitmap"
PASS :offscreenCanvas,snapshot:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                               1,0,0                                0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                         00 00 00 ff                          00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000     0.00000,0.00000,0.00000,1.00000      0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.380392,1.00000 0.168627,0.392157,0.0980392,1.00000 0.364706,0.0549020,0.0274510,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                           0,0,0,255                            0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                       0,0,97,255                       43,100,25,255                          93,14,7,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                 0,0,-97,0                      -43,-100,-25,0                         -93,-14,-7,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 1,1,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                             1,1,0
      act. texel bytes (little-endian) ==      0x:                       00 00 00 ff
                           act. colors == R,G,B,A:   0.00000,0.00000,0.00000,1.00000
                           exp. colors == R,G,B,A: 0.400000,0.400000,0.00000,1.00000
            act. normal-ULPs-from-zero == R,G,B,A:                         0,0,0,255
            exp. normal-ULPs-from-zero == R,G,B,A:                     102,102,0,255
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                              -102,-102,0,0
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 1,1,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                             1,1,0
      act. texel bytes (little-endian) ==      0x:                       00 00 00 ff
                           act. colors == R,G,B,A:   0.00000,0.00000,0.00000,1.00000
                           exp. colors == R,G,B,A: 0.400000,0.400000,0.00000,1.00000
            act. normal-ULPs-from-zero == R,G,B,A:                         0,0,0,255
            exp. normal-ULPs-from-zero == R,G,B,A:                     102,102,0,255
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                              -102,-102,0,0
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="convertToBlob"
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="transferToImageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,1.00000,0.400000 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,255,102                      0,255,0,102                      255,0,0,102 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,0,-255,-102                    0,-255,0,-102                    -255,0,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,1.00000,0.400000 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,255,102                      0,255,0,102                      255,0,0,102 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,0,-255,-102                    0,-255,0,-102                    -255,0,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="srgb";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="srgb";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                               1,0,0                                0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                         00 00 00 ff                          00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000     0.00000,0.00000,0.00000,1.00000      0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.380392,1.00000 0.168627,0.392157,0.0980392,1.00000 0.364706,0.0549020,0.0274510,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                           0,0,0,255                            0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                       0,0,97,255                       43,100,25,255                          93,14,7,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                 0,0,-97,0                      -43,-100,-25,0                         -93,-14,-7,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="srgb";snapshotType="imageBitmap"
FAIL :offscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";snapshotType="convertToBlob" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.419608,1.00000 0.00000,0.407843,0.00000,1.00000 0.439216,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,107,255                      0,104,0,255                      112,0,0,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-107,0                       0,-104,0,0                       -112,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
 Reached unreachable code
FAIL :offscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.400000,1.00000 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,102,255                      0,102,0,255                      102,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-102,0                       0,-102,0,0                       -102,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.419608,1.00000 0.00000,0.407843,0.00000,1.00000 0.439216,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,107,255                      0,104,0,255                      112,0,0,255 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-107,0                       0,-104,0,0                       -112,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,1.00000,0.400000 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,255,102                      0,255,0,102                      255,0,0,102 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,0,-255,-102                    0,-255,0,-102                    -255,0,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                             0,0,0                               1,0,0                               0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                       00 00 00 00                         00 00 00 00                         00 00 00 00 ...
                           act. colors == R,G,B,A:   0.00000,0.00000,0.00000,0.00000     0.00000,0.00000,0.00000,0.00000     0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.960784,0.400000 0.462745,0.980392,0.294118,0.400000 0.921569,0.196078,0.137255,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                           0,0,0,0                             0,0,0,0                             0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                       0,0,245,102                      118,250,75,102                       235,50,35,102 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                              0,0,-245,-102                  -118,-250,-75,-102                   -235,-50,-35,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";snapshotType="imageBitmap"
PASS :offscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="convertToBlob"
FAIL :offscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="transferToImageBitmap" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,1.00000,0.400000 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,255,102                      0,255,0,102                      255,0,0,102 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,0,-255,-102                    0,-255,0,-102                    -255,0,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithSameColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:164:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:195:43
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 00                      00 00 00 00                      00 00 00 00 ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000  0.00000,0.00000,0.00000,0.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,1.00000,0.400000 0.00000,1.00000,0.00000,0.400000 1.00000,0.00000,0.00000,0.400000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                          0,0,0,0                          0,0,0,0                          0,0,0,0 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,255,102                      0,255,0,102                      255,0,0,102 ...
             tolerance ± 2 normal-ULPs
       diff (act - exp) in normal-ULPs ==                             0,0,-255,-102                    0,-255,0,-102                    -255,0,0,-102 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    checkImageResultWithDifferentColorSpaceCanvas@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:184:35
    checkImageResult@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:199:48
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:362:19
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:326:11
 Reached unreachable code
PASS :offscreenCanvas,snapshot:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";snapshotType="imageBitmap"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="bgra8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba8unorm";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="onscreen"
FAIL :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="offscreen" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            1,0,0                            0,1,0                             1,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                       00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000   0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.400000,0.00000,1.00000 0.400000,0.00000,0.00000,1.00000 0.400000,0.400000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                         0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,102,0,255                      102,0,0,255                     102,102,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,-102,0,0                       -102,0,0,0                     -102,-102,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:482:35
 Reached unreachable code
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="onscreen"
FAIL :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="offscreen" assert_unreached:
  - EXPECTATION FAILED: Canvas had unexpected contents:
     between 0,0,0 and 1,1,0 inclusive:
                                coords ==   X,Y,Z:                            0,0,0                            1,0,0                            0,1,0 ...
      act. texel bytes (little-endian) ==      0x:                      00 00 00 ff                      00 00 00 ff                      00 00 00 ff ...
                           act. colors == R,G,B,A:  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000  0.00000,0.00000,0.00000,1.00000 ...
                           exp. colors == R,G,B,A: 0.00000,0.00000,0.419608,1.00000 0.00000,0.407843,0.00000,1.00000 0.439216,0.00000,0.00000,1.00000 ...
            act. normal-ULPs-from-zero == R,G,B,A:                        0,0,0,255                        0,0,0,255                        0,0,0,255 ...
            exp. normal-ULPs-from-zero == R,G,B,A:                      0,0,107,255                      0,104,0,255                      112,0,0,255 ...
             tolerance ± 0 normal-ULPs
       diff (act - exp) in normal-ULPs ==                                0,0,-107,0                       0,-104,0,0                       -112,0,0,0 ...
    ErrorWithExtra@http://127.0.0.1:8000/webgpu/common/util/util.js:23:10
    readPixelsFrom2DCanvasAndCompare@http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:236:25
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:482:35
 Reached unreachable code
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="opaque";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="srgb";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="onscreen";canvas2DType="offscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="onscreen"
PASS :drawTo2DCanvas:format="rgba16float";alphaMode="premultiplied";colorSpace="display-p3";webgpuCanvasType="offscreen";canvas2DType="offscreen"
PASS :transferToImageBitmap_unconfigured_nonzero_size:readbackCanvasType="onscreen"
PASS :transferToImageBitmap_unconfigured_nonzero_size:readbackCanvasType="offscreen"
FAIL :transferToImageBitmap_zero_size:configure=true assert_unreached:
  - EXPECTATION FAILED: unexpectedly did not throw
    shouldThrow@http://127.0.0.1:8000/webgpu/common/framework/fixture.js:321:45
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:544:16
 Reached unreachable code
FAIL :transferToImageBitmap_zero_size:configure=false assert_unreached:
  - EXPECTATION FAILED: unexpectedly did not throw
    shouldThrow@http://127.0.0.1:8000/webgpu/common/framework/fixture.js:321:45
    @http://127.0.0.1:8000/webgpu/webgpu/web_platform/canvas/readbackFromWebGPUCanvas.spec.js:544:16
 Reached unreachable code
PASS :transferToImageBitmap_huge_size:
